Add files via upload
Computation of Flag Dressian and positive Flag Dressian using Polymake. That is, the set of points that induce subdivisions of the 4-permutahedron into cells that are generalized permutahedra, and Bruhat interval poytopes respectively. If you want these fans, but don't want to perform the somewhat time consuming computations in this notebook, they are available as polymake objects in this repository.

This is a demonstration of the main theorems in https://arxiv.org/pdf/2111.13676.pdf